PPMD: fix three defects the allocator harness could not previously reach
The suballocator harness ran 48 minutes in CI and was killed with nothing
printed. It was not a CI problem, and it reproduces locally: seed 42 on an 8 KB
heap wedges the C driver at operation 146, and a faithful port wedges with it.

GlueFreeBlocks absorbs a following free block with

    while ((p1 = p + p->NU)->Stamp == ~0U) { p->NU += p1->NU;  p1->NU = 0; }

which clears the absorbed block's NU but leaves its Stamp reading ~0U. The heap
therefore accumulates husks: twelve bytes that still say "free" and claim zero
units. When a later block's end lands on one, the loop adds zero, rewrites zero
and recomputes the same p1 for ever. Traced both husks the harness hit back to
the absorb that zeroed them, in the same run. The port now stops there, which
cannot change any output the C can produce -- in that state the C never leaves
the loop and so never returns a byte.

Chasing that exposed two more, both in code the old harness never reached:

  * GetUsedMemory can go negative. BList[i].Stamp counts blocks and this charges
    each Indx2Units[i] units, but GlueFreeBlocks inserts remainders smaller than
    their list's nominal size, so the free-list total can exceed what is free.
    The C absorbs it in DWORD; the port was in usize and panicked. Width matters
    as much as wrapping: the model compares the result against SubAllocatorSize
    on the branch that decides a restart, and a 64-bit wrap gives a different
    answer to a 32-bit one. Seed 1 on a 2 KB heap reaches it in 124 operations.

  * A per-symbol eprintln! left in encode_file. All 798 CI streams stayed
    byte-identical, because the harness sends stderr to /dev/null -- the only
    symptom was a job that took 634 seconds instead of twenty. Measured 470x
    slower than the C on 264 KB of text; 1.2x after removal.

Harness changes, all of them things whose absence hid one of the above:

  * Every driver run is capped (perl, not timeout(1) -- macOS has neither
    timeout nor gtimeout). A C that runs over is reported, then compared again
    at the largest op count it does return from, found by bisection.
  * A 2 KB heap added. 8 KB never exhausted: its unit area is 595 units and this
    op mix allocates about that many, stopping just short. At 2 KB every seed
    fails to allocate 3-5 times, which is what found the GetUsedMemory wrap.
  * Both harnesses print per-combination progress. A silent run and a hung one
    were indistinguishable, which is why the 48-minute job went unnoticed.
  * ppmd-check asserts the drivers are silent on stderr and that the port is
    within 10x of the C, before it starts discarding stderr and stops timing.

CI: the job is no longer marked temporary, its grid widens to 9 orders x 4
budgets, and the allocator step gets its own 10-minute cap so it can never again
take a passing grid down with it. The budgets are all small deliberately --
measured at order 16, the 150 KB noise input gives different output at 1, 2 and
3 MB and byte-identical output at 4, 8 and 16 MB, because without exhaustion the
model never restarts and the budget stops being an input at all.

Locally: 456/456 streams byte-identical over 8 combinations, 16/16 allocator
traces identical, 3/3 seeds exhausting the heap, 2 combinations reaching the
glue loop where the port returns and the C does not.

+++ rust/darc-codecs/src/exports.rs
@@ -1543,7 +1543,7 @@ pub unsafe extern "C" fn darc_rs_ppmd_sa_stop() { sa().stop() }
 /// # Safety
 /// See [`darc_rs_ppmd_sa_start`].
 #[no_mangle]
-pub unsafe extern "C" fn darc_rs_ppmd_sa_used() -> u32 { sa().get_used_memory() as u32 }
+pub unsafe extern "C" fn darc_rs_ppmd_sa_used() -> u32 { sa().get_used_memory() }
 
 /// # Safety
 /// See [`darc_rs_ppmd_sa_start`].
